First part of project completed
Program the app to get inputs from user, process the inputs and changes values depending on selected values. Used the bind() method with predefined values to make the app work. WAS THE MAIN PURSUIT OF THE PROJECT TO USE THE BIND METHOD.

Understand how it works in small project, and waiting to use it again.